Update predict.py to support more argument

+++ experiment/vivos/predict.py
@@ -4,22 +4,26 @@ import argparse
 parser = argparse.ArgumentParser(description='Process some integers.')
 parser.add_argument('--wav', help='Path for input file to predict', required=True)
 parser.add_argument('--kaldi_folder', help='Kaldi dir path', required=True)
-parser.add_argument('--model_path', help='Model path (default: uts_{number} in kaldi-trunk/egs)')
+parser.add_argument('--model_path', help='Model path (default: exp/{model} in kaldi-trunk/egs/{result})', required=True)
+parser.add_argument('--utils_path', help='Kaldi utils dir path, usually in super parent directory of model_path')
 
 args = parser.parse_args()
 
 
-def predict(kaldi_folder, wav_file, model_path=None):
+def predict(kaldi_folder, wav_file, model_path, utils_path=None):
     # Model path usually is in etc at kaldi-trunk/egs/uts_{random_int}/exp
-    if model_path is not None:
-        model = model_path
-    else:
-        egs_path = "{}/egs".format(kaldi_folder)
-        uts_dirs = [d for d in os.listdir(egs_path) if d.startswith("uts_")]
-        if len(uts_dirs) > 0:
-            model = "{}/{}".format(egs_path,uts_dirs[0])
-        else:
-            raise Exception("Cannot find pre-trained model in kaldi/egs folder.")
+    model = model_path
+
+    if not os.path.exists(os.path.join(model,"final.mdl")):
+        raise Exception("Cannot find final.mdl model file with given model path.")
+    if not os.path.exists(os.path.join(model, "graph")):
+        raise Exception("Cannot find graph with given model path.")
+
+    if utils_path is None:
+        utils_path = os.path.join(os.path.dirname(os.path.dirname(model)), "utils")
+
+    if not os.path.exists(os.path.join(utils_path,"int2sym.pl")):
+        raise Exception("Cannot find int2sym.pl file with given utils path, please make sure that you are provided correctly utils_path argument")
 
     # Prepare predict dir
     os.system("cd {}; rm -rf predict;".format(model))
@@ -30,13 +34,13 @@ def predict(kaldi_folder, wav_file, model_path=None):
     os.system("cd {}/predict/experiment; mkdir triphones_deldel;".format(model))
 
     # Copy pre-trained model
-    os.system("cd {};cp exp/tri2a/35.mdl predict/experiment/triphones_deldel/final.mdl;".format(model))
-    os.system("cd {};cp -r exp/tri2a/graph predict/experiment/triphones_deldel/graph".format(model))
+    os.system("cd {};cp final.mdl predict/experiment/triphones_deldel/final.mdl;".format(model))
+    os.system("cd {};cp -r graph predict/experiment/triphones_deldel/graph".format(model))
     os.system("cd {}/predict/config; echo '--sample-frequency=16000 \
             \n--num-mel-bins=40 \n--frame-length=25 \n--frame-shift=10 \
             \n--high-freq=0 \n--low-freq=0 \n--num-ceps=13 \n--window-type=hamming \
             \n--use-energy=true' > mfcc.conf".format(model))
-    os.system("cd {}/predict/transcriptions; echo 'test {}' > wav.scp".format(model, wav_file))
+    os.system("cd {}/predict/transcriptions; echo 'result: {}' > wav.scp".format(model, wav_file))
     print("done")
 
     # Run predict
@@ -56,9 +60,15 @@ def predict(kaldi_folder, wav_file, model_path=None):
                       --word-symbol-table=experiment/triphones_deldel/graph/words.txt \
                       ark:transcriptions/lattices.ark ark,t:transcriptions/one-best.tra" \
               .format(model, kaldi_folder))
-    os.system("cd {}/predict; {}/utils/int2sym.pl -f 2- {}/predict/experiment/triphones_deldel/graph/words.txt transcriptions/one-best.tra \
+    os.system("cd {}/predict; {}/int2sym.pl -f 2- {}/predict/experiment/triphones_deldel/graph/words.txt transcriptions/one-best.tra \
                       > {}/predict/transcriptions/one-best-hypothesis.txt; echo $(<{}/predict/transcriptions/one-best-hypothesis.txt);" \
-              .format(model, model, model, model, model))
+              .format(model, utils_path, model, model, model))
+
+    result = open("{}/predict/transcriptions/one-best-hypothesis.txt".format(model)).read()
+    #Result will stored in model_path/predict/transcriptions/one-best-hypothesis.txt under format test {predict_result}
+    result = result[8:]
+    print(result)
+    return result
 
 if __name__ == "__main__":
-    predict(args.kaldi_folder, args.wav, args.model_path)
+    predict(args.kaldi_folder, args.wav,args.model_path,args.utils_path)
